Replace linkAddrCache with neighborCache
This change replaces the current implementation of ARP through linkAddrCache,
with an implementation of NUD through neighborCache. ARP is limited. Switching
to using NUD for both ARP and NDP is beneficial for the reasons described by
RFC 4861 Section 3.1:

  "[Using NUD] significantly improves the robustness of packet delivery in the
  presence of failing routers, partially failing or partitioned links, or nodes
  that change their link-layer addresses. For instance, mobile nodes can move
  off-link without losing any connectivity due to stale ARP caches."

  "Unlike ARP, Neighbor Unreachability Detection detects half-link failures and
  avoids sending traffic to neighbors with which two-way connectivity is
  absent."

Along with these changes exposes the API for querying and operating the
neighbor cache. Operations include:
  - Create a static entry
  - List all entries
  - Delete all entries
  - Remove an entry by address

This also exposes the API to change the NUD protocol constants on a per-NIC
basis to allow Neighbor Discovery to operate over links with widely varying
performance characteristics. See [RFC 4861 Section 10][1] for the list of
constants.

Finally, an API is exposed to allow users to subscribe to NUD state changes.
See [RFC 4861 Appendix C][3] for the list of edges.

Tests:
 //third_party/gvisor/pkg/tcpip/network/arp:arp_test
 - TestDirectRequest

 //third_party/gvisor/pkg/tcpip/network/ipv6:ipv6_test
 - TestLinkResolution
 - TestNDPValidation
 - TestNeighorAdvertisementWithTargetLinkLayerOption
 - TestNeighorSolicitationResponse
 - TestNeighorSolicitationWithSourceLinkLayerOption
 - TestRouterAdvertValidation

 //third_party/gvisor/pkg/tcpip/stack:stack_test
 - TestForwardingWithFakeResolver
 - TestForwardingWithFakeResolverManyPackets
 - TestForwardingWithFakeResolverManyResolutions
 - TestForwardingWithFakeResolverPartialTimeout
 - TestForwardingWithFakeResolverTwoPackets
 - TestIPv6SourceAddressSelectionScopeAndSameAddress

[1]: https://tools.ietf.org/html/rfc4861#section-10
[2]: https://tools.ietf.org/html/rfc4861#appendix-C

Fixes #1889
Fixes #1894
Fixes #1895
Fixes #1947
Fixes #1948
Fixes #1949
Fixes #1950

PiperOrigin-RevId: 316767757